move is_fuchsian into PlainDifferentialOperator
