+++ src/ore_algebra/analytic/factorization.py
@@ -28,7 +28,7 @@ from ore_algebra.analytic.monodromy import _monodromy_matrices
 from ore_algebra.analytic.utilities import as_embedded_number_field_elements
 from ore_algebra.guessing import guess
 from sage.arith.functions import lcm
-from sage.arith.misc import valuation, algdep, gcd
+from sage.arith.misc import algdep, gcd
 from sage.functions.all import log, floor
 from sage.functions.other import binomial, factorial
 from sage.matrix.constructor import matrix
@@ -73,32 +73,6 @@ class LinearDifferentialOperator(PlainDifferentialOperator):
         self.precision = 100 #100*self.algebraicity_degree
 
         self.monodromy_data = MonoData(0, [], None, 0)
-
-        self.fuchsian_info = None
-
-
-    def is_fuchsian(self):
-
-        r"""
-        Return True if ``self`` is fuchsian, False otherwise.
-
-        Fuch's criterion: p is a regular point of a_n*Dz^n + ... + a_0 (with a_i
-        polynomial) iff no (z-p)^{n-k}*a_k/a_n admits p as pole.
-        """
-
-        coeffs = self.coefficients()
-        fac = coeffs.pop().factor()
-        for (f, m) in fac:
-            for k, ak in enumerate(coeffs):
-                mk = valuation(ak, f)
-                if mk - m < k - self.order(): return False
-
-        dop = self.annihilator_of_composition(1/self.base_ring().gen())
-        for k, frac in enumerate(dop.monic().coefficients()[:-1]):
-            d = (self.base_ring().gen()**(self.order() - k)*frac).denominator()
-            if d(0)==0: return False
-
-        return True
 
 
     def monodromy(self, precision, verbose=False):
